json rpc server extracted from the language server
