+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/JsonRpc.scala
@@ -1,0 +1,35 @@
+package org.enso.languageserver.protocol
+
+import io.circe.generic.auto._
+import org.enso.jsonrpc.Protocol
+import org.enso.languageserver.capability.CapabilityApi.{
+  AcquireCapability,
+  ForceReleaseCapability,
+  GrantCapability,
+  ReleaseCapability
+}
+import org.enso.languageserver.filemanager.FileManagerApi._
+import org.enso.languageserver.text.TextApi._
+
+object JsonRpc {
+
+  val protocol: Protocol = Protocol.empty
+    .registerRequest(AcquireCapability)
+    .registerRequest(ReleaseCapability)
+    .registerRequest(WriteFile)
+    .registerRequest(ReadFile)
+    .registerRequest(CreateFile)
+    .registerRequest(OpenFile)
+    .registerRequest(CloseFile)
+    .registerRequest(SaveFile)
+    .registerRequest(ApplyEdit)
+    .registerRequest(DeleteFile)
+    .registerRequest(CopyFile)
+    .registerRequest(MoveFile)
+    .registerRequest(ExistsFile)
+    .registerRequest(TreeFile)
+    .registerNotification(ForceReleaseCapability)
+    .registerNotification(GrantCapability)
+    .registerNotification(TextDidChange)
+
+}
