+++ engine/language-server/src/main/scala/org/enso/languageserver/requesthandler/ApplyEditHandler.scala
@@ -1,24 +1,12 @@
 package org.enso.languageserver.requesthandler
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Props}
+import org.enso.jsonrpc.Errors.ServiceError
+import org.enso.jsonrpc._
 import org.enso.languageserver.data.Client
-import org.enso.languageserver.jsonrpc.Errors.ServiceError
-import org.enso.languageserver.jsonrpc._
-import org.enso.languageserver.text.TextApi.{
-  ApplyEdit,
-  FileNotOpenedError,
-  InvalidVersionError,
-  TextEditValidationError,
-  WriteDeniedError
-}
+import org.enso.languageserver.text.TextApi._
 import org.enso.languageserver.text.TextProtocol
-import org.enso.languageserver.text.TextProtocol.{
-  ApplyEditSuccess,
-  FileNotOpened,
-  TextEditInvalidVersion,
-  TextEditValidationFailed,
-  WriteDenied
-}
+import org.enso.languageserver.text.TextProtocol.{ApplyEdit => _, _}
 
 import scala.concurrent.duration.FiniteDuration
